+++ framework/Furion/App/App.cs
@@ -50,7 +50,7 @@ namespace Furion
         /// <summary>
         /// 全局配置选项
         /// </summary>
-        public static readonly IConfiguration Configuration;
+        public static IConfiguration Configuration { get; internal set; }
 
         /// <summary>
         /// 获取Web主机环境，如，是否是开发环境，生产环境等